requirements.txt
# ── Core Web Framework ───────────────────────────────────────────────────────
streamlit==1.43.2
streamlit-audiorecorder==0.0.6
nest-asyncio==1.6.0
watchdog==6.0.0

# ── LangChain / LangGraph ────────────────────────────────────────────────────
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.83
langchain-nvidia-ai-endpoints==0.3.9
langchain-openai==0.3.35
langchain-google-genai==2.1.1
langchain-text-splitters==0.3.7
langgraph==0.2.76
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.7
langgraph-sdk==0.1.74
langsmith==0.3.45

# ── NVIDIA AI / Riva (Voice / Whisper) ───────────────────────────────────────
nvidia-riva-client==2.24.0
grpcio==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.71.0
websockets==15.0.1

# ── OpenAI / Inference ───────────────────────────────────────────────────────
openai==2.15.0

# ── Vector DB & Retrieval ────────────────────────────────────────────────────
faiss-cpu==1.10.0
sentence-transformers==5.0.0
rank-bm25==0.2.2

# ── Document Processing ──────────────────────────────────────────────────────
PyMuPDF==1.26.7
PyPDF2==3.0.1
python-docx==1.2.0
python-pptx==1.0.2
openpyxl==3.1.5
xlsxwriter==3.2.9
lxml==6.0.2
filetype==1.2.0

# ── Audio Processing ─────────────────────────────────────────────────────────
pydub==0.25.1

# ── ML / Embeddings ──────────────────────────────────────────────────────────
torch==2.7.1
transformers==4.54.0
accelerate==1.12.0
tokenizers==0.21.4
safetensors==0.5.3
sentencepiece==0.2.1
huggingface-hub==0.34.2
scikit-learn==1.7.1
scipy==1.16.1
numpy==2.2.4

# ── Data / Evaluation ────────────────────────────────────────────────────────
ragas==0.4.3
datasets==4.5.0
pandas==2.2.3
pyarrow==23.0.0

# ── Google APIs ──────────────────────────────────────────────────────────────
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
proto-plus==1.26.1

# ── HTTP / Networking ────────────────────────────────────────────────────────
requests==2.32.3
requests-toolbelt==1.0.0
httpx==0.28.1
httpx-sse==0.4.0
aiohttp==3.11.14
certifi==2025.1.31
urllib3==2.3.0

# ── Serialization / Data ─────────────────────────────────────────────────────
protobuf==5.29.4
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
instructor==1.14.4
dataclasses-json==0.6.7
marshmallow==3.26.1
orjson==3.10.16
PyYAML==6.0.2

# ── Database / Storage ───────────────────────────────────────────────────────
SQLAlchemy==2.0.39
diskcache==5.6.3

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.1.0
tiktoken==0.12.0
tenacity==9.0.0
tqdm==4.67.1
joblib==1.5.1
psutil==7.2.0
packaging==24.2
pillow==11.1.0
regex==2024.11.6
sympy==1.14.0
rich==14.3.1
click==8.1.8
typer==0.21.1
tabulate==0.9.0
six==1.17.0
pytz==2025.2
tzdata==2025.2
python-dateutil==2.9.0.post0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest==9.0.2
fastapi
uvicorn
python-multipart
